+++ statistics/handle/BUILD.bazel
@@ -30,6 +30,7 @@ go_library(
         "//sessionctx/variable",
         "//sessiontxn",
         "//statistics",
+        "//statistics/handle/autoanalyze",
         "//statistics/handle/cache",
         "//statistics/handle/globalstats",
         "//statistics/handle/lockstats",
@@ -43,7 +44,6 @@ go_library(
         "//util/logutil",
         "//util/mathutil",
         "//util/sqlexec",
-        "//util/timeutil",
         "@com_github_klauspost_compress//gzip",
         "@com_github_ngaut_pools//:pools",
         "@com_github_pingcap_errors//:errors",
