+++ statistics/handle/autoanalyze/BUILD.bazel
@@ -1,0 +1,42 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "autoanalyze",
+    srcs = ["autoanalyze.go"],
+    importpath = "github.com/pingcap/tidb/statistics/handle/autoanalyze",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//infoschema",
+        "//kv",
+        "//metrics",
+        "//parser/ast",
+        "//parser/model",
+        "//sessionctx",
+        "//sessionctx/variable",
+        "//statistics",
+        "//util",
+        "//util/chunk",
+        "//util/logutil",
+        "//util/sqlexec",
+        "//util/timeutil",
+        "@com_github_pingcap_errors//:errors",
+        "@org_uber_go_zap//:zap",
+    ],
+)
+
+go_test(
+    name = "autoanalyze_test",
+    timeout = "short",
+    srcs = ["autoanalyze_test.go"],
+    flaky = True,
+    shard_count = 3,
+    deps = [
+        ":autoanalyze",
+        "//parser/model",
+        "//statistics",
+        "//statistics/handle",
+        "//testkit",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//oracle",
+    ],
+)
