+++ statistics/handle/updatetest/BUILD.bazel
@@ -8,7 +8,7 @@ go_test(
         "update_test.go",
     ],
     flaky = True,
-    shard_count = 28,
+    shard_count = 25,
     deps = [
         "//parser/model",
         "//parser/mysql",
@@ -17,6 +17,7 @@ go_test(
         "//sessionctx/variable",
         "//statistics",
         "//statistics/handle",
+        "//statistics/handle/autoanalyze",
         "//testkit",
         "//testkit/testsetup",
         "//types",
@@ -24,7 +25,6 @@ go_test(
         "//util/ranger",
         "@com_github_prometheus_client_model//go",
         "@com_github_stretchr_testify//require",
-        "@com_github_tikv_client_go_v2//oracle",
         "@org_uber_go_goleak//:goleak",
     ],
 )
