+++ statistics/handle/usage/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//kv",
+        "//sessionctx",
         "//types",
         "//util/sqlexec",
         "@com_github_pingcap_errors//:errors",
